The Swedish Arthroplasty Registert theme

Implementation of graphical profile used by the annual report and more.

theme_slr(
  axis_text_angle = NULL,
  legend.position = "bottom",
  legend.justification = legend.position,
  legend_title = FALSE,
  text_size = 7,
  subtitle_size = 8,
  title_size = 9,
  title_hjust = 0.5,
  subtitle = FALSE,
  x_lab_exists = FALSE,
  title_margin = if (subtitle) 1 else title_size/2
)

Arguments

axis_text_angle

passed to axis.text.x = element_text(angle = axis_text_angle)

legend.position

the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)

legend.justification

anchor point for positioning legend inside plot ("center" or two-element numeric vector) or the justification according to the plot area when positioned outside the plot

legend_title

Include legend title (boolean)

text_size, title_size, subtitle_size

Text size for most text, title and subtitle

title_hjust, title_margin

passed to plot.title = element_text(hjust = title_hjust, margin = margin(b = title_margin)

subtitle

is a subtitle used in the figure (boolean)?

x_lab_exists

Indicator if x label should be blank or not

Value

Modified version of theme_classic()